pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "is469-wealth-manager"
version = "0.1.0"
description = "IS469 Wealth Manager — Sentiment Analysis & Portfolio Agents"
requires-python = ">=3.10"
dependencies = [
    # ── Core ML ───────────────────────────────────────────────────────────────
    "torch>=2.1.0",
    "transformers==4.56.2",
    "datasets>=2.19.1",
    "huggingface-hub>=0.34.0",
    "hf-transfer",
    "sentencepiece",
    "protobuf>=4.25.3",
    "accelerate",
    "peft",
    "sentence-transformers>=2.7.0",
    # ── QLoRA / unsloth fine-tuning ───────────────────────────────────────────
    "unsloth",
    "bitsandbytes",
    "trl==0.22.2",
    # ── Evaluation & data science ─────────────────────────────────────────────
    "scikit-learn",
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "matplotlib",
    "tqdm",
    "ragas>=0.1.9",
    # ── Environment / notebook utils ──────────────────────────────────────────
    "python-dotenv",
    "ipywidgets",
    "jupyter",
    # ── LLM orchestration ─────────────────────────────────────────────────────
    "langgraph>=1.1.3",
    "langchain>=0.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=0.1.7",
    "langchain-anthropic>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "anthropic>=0.28.0",
    "mcp>=1.0.0",
    # ── Financial data ────────────────────────────────────────────────────────
    "yfinance>=1.2.0",
    "PyPortfolioOpt>=1.5.5",
    # ── RAG & vector database ─────────────────────────────────────────────────
    "chromadb>=1.4.0",
    "pypdf>=4.2.0",
    "tiktoken>=0.7.0",
    "rank-bm25>=0.2.2",
    # ── Web server ────────────────────────────────────────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.24.0",
]

[tool.uv]
# Pin exact versions for reproducibility where needed
override-dependencies = ["transformers==4.56.2", "trl==0.22.2"]

[tool.hatch.build.targets.wheel]
packages = ["agents", "graph"]

[tool.uv.sources]
# unsloth requires a special index on Colab/CUDA environments;
# locally uv will pull from PyPI (CPU build for dev, GPU on Colab)
